// projects/library/application-service/src/lib/tsc/enum/application.enum.ts
export enum ApplicationKeyID {
    NETZERO_ADMIN_KEY_ID = 'NETZERO_ADMIN',
    ACCOUNT_KEY_ID = 'ACCOUNT',
    ADMIN_KEY_ID = 'ADMIN',
    NETZERO_KEY_ID = 'TSC_NETZERO',
    ENERGY_KEY_ID = 'TSC_ENERGY',
    WATER_KEY_ID = 'TSC_WATER',
    ENVIRO_KEY_ID = 'TSC_ENVIRO',
    CONNECT_KEY_ID = 'TSC_CONNECT',
    OFFSET_KEY_ID = 'TSC_OFFSET',
    SUPPLIER_KEY_ID = 'SUPPLIER',
    EMPLOYEE_KEY_ID = 'EMPLOYEE'
}